@@ -1,14 +1,14 @@
 module github.com/v2fly/domain-list-community
 
-go 1.18
+go 1.19
 
 require (
-	github.com/v2fly/v2ray-core/v5 v5.1.0
-	google.golang.org/protobuf v1.28.1
+	github.com/v2fly/v2ray-core/v5 v5.10.1
+	google.golang.org/protobuf v1.31.0
 )
 
 require (
 	github.com/adrg/xdg v0.4.0 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	golang.org/x/sys v0.13.0 // indirect
 )
